Moved format_docstr fn to output module
