+++ src/lib.rs
@@ -3,6 +3,6 @@
 mod analyzer;
 pub use analyzer::{analyze, Config};
 mod output;
-pub use output::{Container, MapType, Member, Output};
+pub use output::{format_docstr, Container, MapType, Member, Output};
 mod derive;
 pub use derive::Derive;
