+++ src/output.rs
@@ -255,10 +255,35 @@ impl MapType {
     }
 }
 
+pub fn format_docstr(indent: &str, input: &str) -> String {
+    let rustdoc_code = "```\n";
+    let rustdoc_starts: Vec<usize> = input
+        .match_indices(rustdoc_code)
+        .step_by(2)
+        .map(|(index, _)| index)
+        .collect();
+    let mut cleaned_input = String::with_capacity(input.len());
+    let mut start = 0;
+    for doc in rustdoc_starts {
+        cleaned_input.push_str(&input[start..doc]);
+        cleaned_input.push_str("```text\n");
+        start = doc + rustdoc_code.len();
+    }
+    cleaned_input.push_str(&input[start..]);
+
+    // TODO: maybe logic to split doc strings by sentence / length here
+
+    format!(
+        "{}/// {}",
+        indent,
+        cleaned_input.replace('\n', &format!("\n{}/// ", indent))
+    )
+}
+
 // unit tests
 #[cfg(test)]
 mod test {
-    use super::{Container, Member};
+    use super::{format_docstr, Container, Member};
     fn name_only_enum_member(name: &str) -> Member {
         Member {
             name: name.to_string(),
@@ -411,5 +436,29 @@ mod test {
         assert!(containers[5].can_derive_default(&containers)); // ReferencesEnumOption
         assert!(containers[6].can_derive_default(&containers)); // ReferencesEnumVec
         assert!(containers[7].can_derive_default(&containers)); // ReferencesEnumNestedOption
+    }
+
+    #[test]
+    fn escapes_codes_from_descriptions() {
+        assert_eq!(
+            "/// ```text\n/// foobar\n/// ```\n/// ",
+            format_docstr("", "```\nfoobar\n```\n")
+        );
+        assert_eq!(
+            "/// Some docs\n/// ```text\n/// foobar\n/// ```\n/// ",
+            format_docstr("", "Some docs\n```\nfoobar\n```\n")
+        );
+        assert_eq!(
+            "/// Some docs\n/// ```text\n/// foobar\n/// ```",
+            format_docstr("", "Some docs\n```\nfoobar\n```")
+        );
+        assert_eq!(
+            "/// ```text\n/// foobar\n/// ```",
+            format_docstr("", "```\nfoobar\n```")
+        );
+        assert_eq!(
+            "/// Some docs\n/// with no code blocks!",
+            format_docstr("", "Some docs\nwith no code blocks!")
+        );
     }
 }
